+++ example/cli/src/app/app-routing.module.ts
@@ -1,0 +1,32 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { BasicTreeComponent } from './basictree/basictree.component';
+import { FullTreeComponent } from './fulltree/fulltree.component';
+import { TemplatesComponent } from './templates/templates.component';
+import { FilterComponent } from './filter/filter.component';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: FullTreeComponent
+  },
+  {
+    path: 'basic',
+    component: BasicTreeComponent
+  },
+  {
+    path: 'templates',
+    component: TemplatesComponent
+  },
+  {
+    path: 'filter',
+    component: FilterComponent
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forRoot(routes, { useHash: true })],
+  exports: [RouterModule]
+})
+export class AppRoutingModule { }