+++ example/cli/src/app/app.module.ts
@@ -3,32 +3,13 @@ import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 import { TreeModule } from 'angular-tree-component';
-import { Router, Route, RouterModule } from '@angular/router';
 
+import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BasicTreeComponent } from './basictree/basictree.component';
 import { FullTreeComponent } from './fulltree/fulltree.component';
 import { TemplatesComponent } from './templates/templates.component';
 import { FilterComponent } from './filter/filter.component';
-
-const routes: Route[] = [
-  {
-    path: '',
-    component: FullTreeComponent
-  },
-  {
-    path: 'basic',
-    component: BasicTreeComponent
-  },
-  {
-    path: 'templates',
-    component: TemplatesComponent
-  },
-  {
-    path: 'filter',
-    component: FilterComponent
-  }
-];
 
 @NgModule({
   declarations: [
@@ -43,7 +24,7 @@ const routes: Route[] = [
     FormsModule,
     HttpModule,
     TreeModule,
-    RouterModule.forRoot(routes, { useHash: true })
+    AppRoutingModule
   ],
   providers: [],
   bootstrap: [AppComponent]