+++ example/cli/src/app/templates/templates.component.ts
@@ -9,18 +9,18 @@ import { ITreeOptions } from 'angular-tree-component';
 
     <h3>treeNodeTemplate and loadingTemplate</h3>
     <tree-root id="tree1" [focused]="true" [nodes]="nodes1" [options]="options1">
-      <template #treeNodeTemplate let-node let-index="index">
+      <ng-template #treeNodeTemplate let-node let-index="index">
         <span [class]="node.data.className + 'Index'">{{ index }}</span>
         <span [class]="node.data.className" [class.title]="true">{{ node.data.title }}</span>
-      </template>
-      <template #loadingTemplate let-node>
+      </ng-template>
+      <ng-template #loadingTemplate let-node>
         <div [class]="node.data.className + 'Loading'">Loading {{ node.data.title }}...</div>
-      </template>
+      </ng-template>
     </tree-root>
 
     <h3>treeNodeFullTemplate</h3>
     <tree-root id="tree2" [focused]="true" [nodes]="nodes2">
-      <template #treeNodeFullTemplate let-node let-index="index">
+      <ng-template #treeNodeFullTemplate let-node let-index="index" let-templates="templates">
         <div
           [class.tree-node]="true"
           [class.tree-node-expanded]="node.isExpanded && node.hasChildren"
@@ -37,7 +37,7 @@ import { ITreeOptions } from 'angular-tree-component';
           </div>
           <tree-node-children [node]="node" [templates]="templates"></tree-node-children>
         </div>
-      </template>
+      </ng-template>
     </tree-root>
   `,
   styles: []
@@ -62,7 +62,10 @@ export class TemplatesComponent {
     },
     {
       title: 'root2',
-      className: 'root2Class'
+      className: 'root2Class',
+      children: [
+        {title: 'child1', className: 'child1Class'}
+      ]
     }
   ];