+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Passing node in context to loading template
 * Fixed using treeNodeFullTemplate
 * Added more demos
+* Fixed filtering without virtual scroll option
 
 <a name="3.2.3"></a>
 # 3.2.3 (2017-03-08)