+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fixed using treeNodeFullTemplate
 * Added more demos
 * Fixed filtering without virtual scroll option
+* allowDrag allows functions
 
 <a name="3.2.3"></a>
 # 3.2.3 (2017-03-08)